improve adding and updating recipes
Do not require setting a category or uploading an image when
updating a recipe.
A user who adds/updates a recipe isn't supposed to rate it.
If uploading an image while updating a recipe, give a hint to
the browser that we expect an image.
Rename the form button when adding a recipe.